main function finally works

+++ hw4/headers/user.h
@@ -14,12 +14,13 @@ public:
     int id = -1;
     std::string name;
     double rating;
-    int ratingSum;
-    int numberOfRides;
+    double ratingSum;
+    double numberOfRides;
     std::string login;
     int password;
 
     void updateRating(int r);
+
     user(std::string name = "NO_USER", std::string login = "NO_USER", std::string password = "");
 };
